# perf(ftp): reduce number of ftp commands run in install preparation