When connecting to an FTP server not on the same network, the latency for every FTP command being run greatly increases, slowing down the install process, which SMM frequently triggers.

Testing this using SirDigby's server, the latency was around 0.3s, and I observed 20 seconds being spent before even starting to extract mods. This PR reduces that time to ~5 seconds.

Improvements made:
* ftp.Exists - consider the error codes returned by MLST or LIST to determine whether a path exists
    * previously it would check for every path segment whether it exists, resulting in longer wait times the deeper the server is located
    * improved install by ~10s
* ftp.MkDir - find last existing segment by checking up rather than down
    * it is very likely that only one or two directories have to be created, so only a couple of path existence checks will be made
    * improved install by ~1s
    * also greatly improves mod extraction (not benchmarked, but previously the throughput would drop to 100B/s average at time, since a lot of directories have to be created)
* parallelize mod removal
    * mod extraction was done in parallel, so there's no reason why deletion shouldn't too
    * one mod removal takes around 6s (not included in the 20s of the benchmark)
* i.Validate - parallelize
    * it is not performed concurrently with any other FTP operation, so it has the whole connection pool available
    * improved install by ~2.5s
* fewer Validate and GetPlatform calls - obtain platform once and pass it to other function calls
    * Validate was being called 6 times, and GetPlatform 4 times
    * improved install by ~1.5s

Additionally, `.` and `..` are ignored in `ReadDir` results, as non-standard FTP servers that use `LIST -a` would return those too, leaving the possibility of the game dir being deleted if it contained a `.smm` file